+++ packages/linux_secret_store/freedesktop_secret/lib/src/models/create_item_result.dart
@@ -4,6 +4,13 @@ import 'package:freedesktop_secret/src/models/prompt_result.dart';
 final class CreateItemResult extends PromptResult {
   CreateItemResult({required this.item, required super.prompt});
 
+  factory CreateItemResult.fromDBus(List<DBusValue> raw) {
+    return CreateItemResult(
+      item: raw[0].asObjectPath(),
+      prompt: raw[1].asObjectPath(),
+    );
+  }
+
   /// The item created, or the special value '/' if a prompt is necessary.
   final DBusObjectPath item;