+++ packages/linux_secret_store/freedesktop_secret/lib/src/models/open_session_result.dart
@@ -3,6 +3,10 @@ import 'package:dbus/dbus.dart';
 final class OpenSecretSession {
   const OpenSecretSession({required this.output, required this.objectPath});
 
+  factory OpenSecretSession.fromDBus(List<DBusValue> raw) {
+    return OpenSecretSession(output: raw[0], objectPath: raw[1].asObjectPath());
+  }
+
   /// Output of the session algorithm negotiation.
   final DBusValue output;