+++ packages/linux_secret_store/freedesktop_secret/lib/src/models/unlock_result.dart
@@ -4,6 +4,13 @@ import 'package:freedesktop_secret/src/models/prompt_result.dart';
 final class UnlockResult extends PromptResult {
   const UnlockResult({required this.unlocked, required super.prompt});
 
+  factory UnlockResult.fromDBus(List<DBusValue> raw) {
+    return UnlockResult(
+      unlocked: .unmodifiable(raw[0].asObjectPathArray()),
+      prompt: raw[1].asObjectPath(),
+    );
+  }
+
   /// Objects that were unlocked without a prompt.
   final List<DBusObjectPath> unlocked;